Extract shared functionality to getTaxCalculatorExpression
